+++ arduino/STM32/Nucleo/test_servo/test_servo.ino
@@ -36,19 +36,19 @@ void setup()
 
 void loop()
 {
-  for (double pos = POSITION_NEUTRAL; pos <= POSITION_MAX; pos += 0.015)
+  for (float pos = POSITION_NEUTRAL; pos <= POSITION_MAX; pos += 0.015F)
   {
     outp = pos;
     delay(20);
   }
 
-  for (double pos = POSITION_MAX; pos >= POSITION_MIN; pos -= 0.015)
+  for (float pos = POSITION_MAX; pos >= POSITION_MIN; pos -= 0.015F)
   {
     outp = pos;
     delay(20);
   }
   
-  for (double pos = POSITION_MIN; pos <= POSITION_NEUTRAL; pos += 0.015)
+  for (float pos = POSITION_MIN; pos <= POSITION_NEUTRAL; pos += 0.015F)
   {
     outp = pos;
     delay(20);
